phase-two/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
email-validator>=2.1.0
python-multipart>=0.0.9
httpx>=0.26.0
pytest>=8.0.0
